css/css-fonts/parsing/font-stretch-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Level 3: parsing font-stretch with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-stretch-prop">
<meta name="assert" content="font-stretch supports only the grammar 'normal | ultra-condensed | extra-condensed | condensed | semi-condensed | semi-expanded | expanded | extra-expanded | ultra-expanded'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("font-stretch", "ultra-normal");
test_invalid_value("font-stretch", "expanded condensed");
</script>
</body>
</html>
